fixed logging for more than 1000 bytes

+++ teensypid/src/feedforward.cpp
@@ -42,7 +42,7 @@ FeedForwardHelper::FeedForwardHelper(uint8_t calibrationBits, uint8_t feedForwar
 
 void FeedForwardHelper::setReadings(uint16_t* readings, int32_t length)
 {
-    readingsSize = min((uint32_t)length, MAX_ARRAY_SIZE);
+    readingsSize = min(length, MAX_ARRAY_SIZE);
     for (uint32_t i = 0; i < readingsSize; i++)
     {
         this->readings[i] = readings[i];
